+++ src/App.tsx
@@ -10,28 +10,25 @@ import {
   PatientRecordSearch,
   CTWProvider,
   PatientAllergies,
-  PatientConditions,
-  PatientConditionsOutside,
   PatientDiagnosticReports,
   PatientDocuments,
   PatientEncounters,
   PatientImmunizations,
-  PatientMedications,
   PatientMedicationsAll,
-  PatientMedicationsOutside,
   PatientProvider,
   PatientSearch,
   PatientTimeline,
   ZusAggregatedProfile,
   ZusAggregatedProfileIFrame,
   PatientConditionsAll,
+  PatientsTableFQS,
 } from ".";
 
 import { PatientMedicationDispense } from "./components/content/medication-dispense/patient-medication-dispense";
-import { ADTAlertsTable } from "./components/content/patients-adt/patients-adt-alerts-table";
 import { ThemeProviderProps } from "@/components/core/providers/theme/theme-provider";
 import { UseQueryResult } from "@tanstack/react-query";
 import { MatchedPatients } from "./components/content/matched-patients/matched-patients";
+import { PatientsTable } from "./components/content/patients/patients-table-ods";
 
 const {
   VITE_AUTH0_AUDIENCE,
@@ -129,6 +126,18 @@ const components: DemoComponent[] = [
     },
   },
   {
+    name: "patients-ods", // uses ODS
+    render: () => {
+      return <PatientsTable handleRowClick={() => {}} title="Patients Table (ODS)" />;
+    },
+  },
+  {
+    name: "patients-fqs",
+    render: () => {
+      return <PatientsTableFQS handleRowClick={() => {}} title="Patients Table (FQS)" />;
+    },
+  },
+  {
     name: "zap",
     render: () => {
       return <ZusAggregatedProfile title="ZAP" includePatientDemographicsForm={false} />;
@@ -172,7 +181,7 @@ const DemoApp = ({ accessToken = "" }) => (
   >
     <PatientProvider patientID={VITE_PATIENT_ID} systemURL={VITE_SYSTEM_URL}>
       {demoComponents.map((demo, index) => (
-        <div className="ctw-space-y-5 ctw-bg-white ctw-p-1" key={index}>
+        <div className="ctw-space-y-5 ctw-bg-white" key={index}>
           <h3>
             {demo.title} <small>{demo.note}</small>
           </h3>