+++ src/components/content/medications/patient-medications-all.tsx
@@ -51,7 +51,7 @@ function PatientMedicationsAllComponent({
   const rowActions = useRowActions(onAddToRecord);
 
   const openDetails = useMedicationDetailsDrawer({
-    rowActions,
+    rowActions: readOnly ? undefined : rowActions,
     enableDismissAndReadActions: true,
   });
 
@@ -102,7 +102,7 @@ function useRowActions(onAddToRecord?: (record: MedicationStatementModel) => voi
   const { toggleRead } = useToggleRead();
 
   return (record: MedicationStatementModel): RowActionsConfigProp<MedicationStatementModel> =>
-    !record.ownedByBuilder(userBuilderId)
+    record.ownedByBuilder(userBuilderId)
       ? []
       : [
           {