+++ src/components/content/patient-history/patient-history-table.tsx
@@ -4,7 +4,7 @@ import cx from "classnames";
 import { useEffect, useState } from "react";
 import { defaultPatientHistoryFilters, patientHistoryFilters } from "./helpers/filters";
 import { useBuilderPatientHistoryList } from "./use-builder-patient-history-list";
-import { TableOptionProps } from "../patients/patients-table";
+import { TableOptionProps } from "../patients/patients-table-helper";
 import { ResourceTableActions } from "../resource/resource-table-actions";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { FilterChangeEvent } from "@/components/core/filter-bar/filter-bar-types";
@@ -42,6 +42,7 @@ export const PatientHistoryTable = withErrorBoundary(
 
     const {
       data: { patients: responsePatients, total: responseTotal, hasNext } = {},
+      isLoading,
       isFetching,
       isError,
     } = useBuilderPatientHistoryList(pageSize, currentPage - 1, status, excludeFutureJobs);
@@ -96,6 +97,7 @@ export const PatientHistoryTable = withErrorBoundary(
               columns={columns}
               pageSize={pageSize}
               handleRowClick={handleRowClick}
+              isLoading={isLoading}
               hidePagination
             >
               <SimplePagination