+++ src/components/content/patients-adt/helpers/filters.ts
@@ -1,4 +1,4 @@
-import { dismissFilter } from "../resource/filters";
+import { dismissFilter } from "../../resource/filters";
 import { FilterChangeEvent, FilterItem } from "@/components/core/filter-bar/filter-bar-types";
 import { EncounterModel } from "@/fhir/models";