+++ src/components/content/patients-adt/patients-adt-alerts-table.tsx
@@ -1,12 +1,15 @@
 import type { TableColumn } from "@/components/core/table/table-helpers";
 import type { PatientModel } from "@/fhir/models/patient";
+import { faFileLines } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import cx from "classnames";
-import { useState } from "react";
-import { adtFilter, defaultADTFilters } from "./filters";
-import { useADTAlertDetailsDrawer } from "./modal-hooks";
+import { useEffect, useState } from "react";
+import { adtFilter, defaultADTFilters } from "./helpers/filters";
+import { useADTAlertDetailsDrawer } from "./helpers/modal-hooks";
+import { mapEncountersAndNotes, RelatedEncounterMap } from "./helpers/related-encounters";
 import { dedupeAndMergeEncounters } from "../encounters/helpers/filters";
 import { defaultEncounterSort, encounterSortOptions } from "../encounters/helpers/sorts";
-import { TableOptionProps } from "../patients/patients-table";
+import { TableOptionProps } from "../patients/patients-table-helper";
 import { getDateRangeView } from "../resource/helpers/view-date-range";
 import { ResourceTable } from "../resource/resource-table";
 import { ResourceTableActions } from "../resource/resource-table-actions";
@@ -14,29 +17,12 @@ import { EmptyTableNoneFound } from "@/components/core/empty-table";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { Pagination } from "@/components/core/pagination/pagination";
 import { AnalyticsProvider } from "@/components/core/providers/analytics/analytics-provider";
-import { CTWRequestContext } from "@/components/core/providers/ctw-context";
 import { useCTW } from "@/components/core/providers/use-ctw";
 import { SimpleMoreList } from "@/components/core/simple-more-list";
 import { EncounterModel } from "@/fhir/models/encounter";
 import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
-import { createGraphqlClient, fqsRequest } from "@/services/fqs/client";
-import { EncounterGraphqlResponse, encountersQuery } from "@/services/fqs/queries/encounters";
-import { QUERY_KEY_ENCOUNTERS_RELATED } from "@/utils/query-keys";
-import { queryClient } from "@/utils/request";
 
 const PAGE_SIZE = 10;
-
-// Cache related encounters and notes for 10 minutes.
-const RELATED_ENC_STALE_TIME = 1000 * 60 * 10;
-
-type RelatedEncounterMap = Map<
-  string,
-  {
-    upid: string;
-    cwcq_encounter_id: string;
-    binary_id: string;
-  }
->;
 
 export type ADTTableProps = {
   className?: cx.Argument;
@@ -45,48 +31,6 @@ export type ADTTableProps = {
   encounterAndNotesData?: RelatedEncounterMap;
   goToPatient: (upid: string) => void;
 } & TableOptionProps<EncounterModel>;
-
-function assignEncountersAndNotes(
-  adtEncounters: EncounterModel[],
-  encounterAndNotesData: RelatedEncounterMap,
-  getRequestContext: () => Promise<CTWRequestContext>
-) {
-  adtEncounters.forEach(async (e: EncounterModel) => {
-    if (!encounterAndNotesData.has(e.resource.id ?? "")) {
-      return;
-    }
-    const encAndNote = encounterAndNotesData.get(e.resource.id ?? "");
-    if (encAndNote) {
-      const { data: encounterFqsData } = await queryClient.fetchQuery(
-        [QUERY_KEY_ENCOUNTERS_RELATED, e.id],
-        async () => {
-          const requestContext = await getRequestContext();
-          const graphClient = createGraphqlClient(requestContext);
-          return fqsRequest<EncounterGraphqlResponse>(graphClient, encountersQuery, {
-            upid: encAndNote.upid,
-            cursor: "",
-            first: 1,
-            sort: {
-              lastUpdated: "DESC",
-            },
-            filter: {
-              ids: {
-                anymatch: [encAndNote.cwcq_encounter_id],
-              },
-            },
-          });
-        },
-        { staleTime: RELATED_ENC_STALE_TIME }
-      );
-
-      const encounterNodes = encounterFqsData.EncounterConnection.edges.map((x) => x.node);
-      const encounterNode = encounterNodes[0];
-      e.relatedEncounter = new EncounterModel(encounterNode, encounterNode.ProvenanceList);
-
-      e.relatedEncounter.binaryId = encAndNote.binary_id.replaceAll('"', "");
-    }
-  });
-}
 
 function ADTTableComponent({
   className,
@@ -97,8 +41,22 @@ function ADTTableComponent({
 }: ADTTableProps) {
   const openADTDetails = useADTAlertDetailsDrawer(goToPatient);
   const [currentPage, setCurrentPage] = useState(1);
+  const [dataEnriched, setDataEnriched] = useState<EncounterModel[]>([]);
+  const [isLoadingDetails, setIsLoadingDetails] = useState(true);
   const { getRequestContext } = useCTW();
-  const { past7days, past30days, past3months } = getDateRangeView<EncounterModel>("periodStart");
+  const { past7days, past30days } = getDateRangeView<EncounterModel>("periodStart");
+
+  useEffect(() => {
+    if (encounterAndNotesData) {
+      void mapEncountersAndNotes(data, encounterAndNotesData, getRequestContext).then(
+        (newEncounters) => {
+          setDataEnriched(newEncounters);
+          setIsLoadingDetails(false);
+        }
+      );
+    }
+  }, [data, encounterAndNotesData, getRequestContext]);
+
   const {
     data: dataFilteredSorted,
     setViewOption,
@@ -108,19 +66,15 @@ function ADTTableComponent({
     defaultView: past30days,
     defaultFilters: defaultADTFilters,
     defaultSort: defaultEncounterSort,
-    records: data,
+    records: dataEnriched,
   });
 
-  const viewOptions = [past7days, past30days, past3months];
+  const viewOptions = [past7days, past30days];
   const dataFilteredSortedDeduped = dedupeAndMergeEncounters(dataFilteredSorted, "patientsADT");
   const dataOnPage = dataFilteredSortedDeduped.slice(
     (currentPage - 1) * PAGE_SIZE,
     currentPage * PAGE_SIZE
   );
-
-  if (encounterAndNotesData) {
-    assignEncountersAndNotes(dataOnPage, encounterAndNotesData, getRequestContext);
-  }
 
   return (
     <AnalyticsProvider componentName="ADTTable">
@@ -145,7 +99,7 @@ function ADTTableComponent({
         <ResourceTable
           data={dataOnPage}
           columns={columns}
-          isLoading={isLoading}
+          isLoading={isLoading || isLoadingDetails}
           emptyMessage={
             <EmptyTableNoneFound
               hasZeroFilteredRecords={dataFilteredSortedDeduped.length === 0}
@@ -154,7 +108,6 @@ function ADTTableComponent({
           }
           onRowClick={openADTDetails}
           hidePagination
-          enableDismissAndReadActions
         >
           <Pagination
             currentPage={currentPage}
@@ -191,6 +144,13 @@ const columns: TableColumn<EncounterModel>[] = [
   {
     title: "Location",
     render: (e) => e.location,
+  },
+  {
+    render: (e) =>
+      e.relatedEncounter?.binaryId ? (
+        <FontAwesomeIcon icon={faFileLines} className="ctw-h-4 ctw-w-4 ctw-text-content-lighter" />
+      ) : undefined,
+    widthPercent: 3,
   },
   {
     title: "Diagnosis",