+++ src/services/fqs/queries/allergies.ts
@@ -21,6 +21,7 @@ export interface AllergyGraphqlResponse {
 export const allergyQuery = gql`
   ${fragmentOrganization}
   ${fragmentCoding}
+  ${fragmentOrganization}
   ${fragmentPatient}
   ${fragmentEncounterReference}
   ${fragmentPractitioner}